+++ photutils/segmentation/deblend.py
@@ -6,6 +6,7 @@ a segmentation image.
 
 import warnings
 
+from astropy.utils.decorators import deprecated_renamed_argument
 from astropy.utils.exceptions import AstropyUserWarning
 import numpy as np
 
@@ -17,6 +18,11 @@ from ..utils._optional_deps import HAS_TQDM  # noqa
 __all__ = ['deblend_sources']
 
 
+@deprecated_renamed_argument('kernel', None, '1.5', message='"kernel" was '
+                             'deprecated in version 1.5 and will be removed '
+                             'in a future version. Instead, if filtering is '
+                             'desired, please input a convolved image '
+                             'directly into the "data" parameter.')
 def deblend_sources(data, segment_img, npixels, kernel=None, labels=None,
                     nlevels=32, contrast=0.001, mode='exponential',
                     connectivity=8, relabel=True, progress_bar=True):
@@ -49,6 +55,9 @@ def deblend_sources(data, segment_img, npixels, kernel=None, labels=None,
         positive integer.
 
     kernel : 2D `~numpy.ndarray` or `~astropy.convolution.Kernel2D`, optional
+        Deprecated. If filtering is desired, please input a convolved
+        image directly into the ``data`` parameter.
+
         The 2D kernel used to filter the image before thresholding.
         Filtering the image will smooth the noise and maximize
         detectability of objects with a shape similar to the kernel.